requirements/base.txt
# Common dependencies shared by local development, validation, and runtime wrappers.
# Install an accelerator-specific PaddlePaddle build separately for OCR execution.
jsonschema>=4.18,<5.0
llama-cpp-python>=0.2.90
paddleocr>=2.7,<3.0
Pillow>=9.5,<12.0
